# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "WEMU API Documentation",
        "description": "Swagger documentation for all APIs",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/docs",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/my-current-subscription": {
            "get": {
                "tags": [
                    "Subscription"
                ],
                "summary": "Get current user subscription",
                "operationId": "468adeb270a9423276450257986b4cc0",
                "responses": {
                    "200": {
                        "description": "Subscription fetched successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/subscription/purchase": {
            "post": {
                "tags": [
                    "Subscription"
                ],
                "summary": "Purchase a subscription",
                "operationId": "e7b7ede197f2d6f9d85121d2db5e95aa",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "subscription_id",
                                    "payment_method_id"
                                ],
                                "properties": {
                                    "subscription_id": {
                                        "type": "integer"
                                    },
                                    "payment_method_id": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Subscription purchased successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}